Removed NuGetPackageUtilsTests and moved the tests to NugetPackageUtilTests.cs. Also, fixed the existing tests to use the new signature for InstallFromSourceAsync
